Crate shardio

Efficiently write Rust structs to shard files from multiple threads. You can process different subsets of your data in different threads, different processes. When reading shardio will merge the data on the fly into a single sorted view.

#[macro_use]
extern crate serde_derive;

extern crate shardio;
use shardio::*;
use std::fs::File;

#[derive(Clone, Eq, PartialEq, Serialize, Deserialize, PartialOrd, Ord, Debug)]
struct Test {
    a: u64,
    b: u32,
}

fn main()
{
    let filename = "test.shardio";
    {
        // Open a shardio output file
        // Parameters here  buffering, and the size of disk chunks
        // which affect how many disk chunks need to be read to
        // satisfy a range query when reading.
        let manager: ShardWriter<Test, Test> =
            ShardWriter::new(filename, 64, 256, 1<<16).unwrap();

        // Get a handle to send data to the file
        let mut sender = manager.get_sender();

        // Generate some test data
        for i in 0..(2 << 16) {
            sender.send(Test { a: (i%25) as u64, b: (i%100) as u32 });
        }
    }

    // Open finished file & test chunked reads
    let reader = ShardReader::<Test, Test>::open(filename);

    // Read back data
    let mut all_items = Vec::new();

    let chunks = reader.make_chunks(5, &Range::all());
    for c in chunks {
        let mut range_iter = reader.iter_range(&c);
        all_items.extend(range_iter);
    }

    // Data will be return in sorted order
    let mut all_items_sorted = all_items.clone();
    all_items.sort();
    assert_eq!(all_items, all_items_sorted);
}

Re-exports

pub use range::Range;

Modules

helper

pmap

range

Structs

DefaultSort
Marker struct for sorting types that implement Ord in their 'natural' order.

MergeIterator
Iterator over merged shardio files

ShardItemIter
Iterator of items from a single shardio reader

ShardReader
Read from a collection of shardio files. The input data is merged to give a single sorted view of the combined dataset. The input files must be created with the same sort order S as they are read with.

ShardSender
A handle that is used to send data to a ShardWriter. Each thread that is producing data needs it's own ShardSender. A ShardSender can be obtained with the get_sender method of ShardWriter. ShardSender implement clone.

ShardWriter
Write a stream data items of type T to disk.

Traits

SortKey
Specify a key function from data items of type T to a sort key of type K. Implment this trait to create a custom sort order.