Struct shardio::ShardReader

pub struct ShardReader<T, K, S = DefaultSort> { /* fields omitted */ }

Read from a collection of shardio files. The input data is merged to give a single sorted view of the combined dataset. The input files must be created with the same sort order S as they are read with.

Methods

impl<T, K, S> ShardReader<T, K, S> where
T: DeserializeOwned,
K: Clone + Ord + DeserializeOwned,
S: SortKey<T, K>,

pub fn open<P: AsRef<Path>>(shard_file: P) -> ShardReader<T, K, S>

Open a single shard files into reader

pub fn open_set<P: AsRef<Path>>(shard_files: &[P]) -> ShardReader<T, K, S>

Open a set of shard files into an aggregated reader

pub fn read_range(&self, range: &Range<K>, data: &mut Vec<T>)

Read data from the given range into data buffer. The data buffer is not cleared before adding items.

Important traits for MergeIterator<'a, T, K, S>
impl<'a, T: 'a, K: 'a, S: 'a> Iterator for MergeIterator<'a, T, K, S> where
K: Clone + Ord,
S: SortKey<T, K>,
ShardItemIter<'a, T, K, S>: Iterator<Item = T>, type Item = T;

pub fn iter_range<'a>(&'a self, range: &Range<K>) -> MergeIterator<'a, T, K, S>

Iterate over items in the given range

Important traits for MergeIterator<'a, T, K, S>
impl<'a, T: 'a, K: 'a, S: 'a> Iterator for MergeIterator<'a, T, K, S> where
K: Clone + Ord,
S: SortKey<T, K>,
ShardItemIter<'a, T, K, S>: Iterator<Item = T>, type Item = T;

pub fn iter<'a>(&'a self) -> MergeIterator<'a, T, K, S>

Iterate over all items

pub fn len(&self) -> usize

Total number of items

pub fn make_chunks(&self, num_chunks: usize, range: &Range<K>) -> Vec<Range<K>>

Generate num_chunks ranges covering the give range, each with a roughly equal numbers of elements. The ranges can be fed to iter_range